insert and update added

+++ src/main/java/edu/gatech/chai/omopv5/jpa/dao/IDao.java
@@ -22,5 +22,5 @@ import java.util.Map;
 
 public interface IDao {
 	public ResultSet runQuery(String query) throws SQLException;
-	public int updateQuery(String query) throws SQLException;
+	public Long updateQuery(String query) throws SQLException;
 }
